working on dialogs and routes

+++ src/composables/useSharedActions.ts
@@ -1,25 +1,24 @@
-import DialogConfirmStrict from '@/components/dialogs/DialogConfirmStrict.vue'
-import DialogInspect from '@/components/dialogs/DialogInspect.vue'
+import useDialogs from '@/composables/useDialogs'
 import useLogger from '@/composables/useLogger'
 import DatabaseManager from '@/services/DatabaseManager'
 import DB from '@/services/db'
 import { deleteIcon } from '@/shared/icons'
 import type { IdType } from '@/shared/types'
-import { useQuasar } from 'quasar'
 
 export default function useSharedActions() {
-    const $q = useQuasar()
     const { log } = useLogger()
+    const { inspectDialog, strictConfirmDialog } = useDialogs()
 
     /**
-     * Fullscreen dialog that provides a human readable view of a model's data.
+     * Opens a fullscreen dialog to inspect a record.
      */
     async function onInspectDialog(id: IdType) {
         const service = DatabaseManager.getService(id)
-        const record = await service.get(DB, id)
-        $q.dialog({
-            component: DialogInspect,
-            componentProps: { record },
+        inspectDialog({
+            labelSingular: service.labelSingular,
+            labelPlural: service.labelPlural,
+            table: service.table,
+            record: await service.get(DB, id),
         })
     }
 
@@ -27,23 +26,20 @@ export default function useSharedActions() {
      * Generic record deletion function. Not for use with SETTINGS or LOGS tables.
      */
     function onDeleteRecord(id: IdType) {
-        $q.dialog({
-            component: DialogConfirmStrict,
-            componentProps: {
-                title: 'Delete Record',
-                message: `Are you sure you want to delete ${id}?`,
-                color: 'negative',
-                icon: deleteIcon,
-                code: 'YES',
+        strictConfirmDialog({
+            title: 'Delete Record',
+            message: `Are you sure you want to delete ${id}?`,
+            color: 'negative',
+            icon: deleteIcon,
+            onOk: async () => {
+                try {
+                    const service = DatabaseManager.getService(id)
+                    const deletedRecord = await service.delete(DB, id)
+                    log.info(`Deleted record`, deletedRecord)
+                } catch (error) {
+                    log.error(`Error deleting record`, error as Error)
+                }
             },
-        }).onOk(async () => {
-            try {
-                const service = DatabaseManager.getService(id)
-                const deletedRecord = await service.delete(DB, id)
-                log.info(`Deleted record`, deletedRecord)
-            } catch (error) {
-                log.error(`Error deleting record`, error as Error)
-            }
         })
     }
